Rename api to assessmentAPI
