+++ src/sagas/index.ts
@@ -6,7 +6,7 @@ import { call, put, race, select, take, takeEvery } from 'redux-saga/effects'
 import * as actions from '../actions'
 import * as actionTypes from '../actions/actionTypes'
 import { WorkspaceLocation } from '../actions/workspaces'
-import { mockAssessmentOverviews, mockAssessments } from '../mocks/api'
+import { mockAssessmentOverviews, mockAssessments } from '../mocks/assessmentAPI'
 import { IState } from '../reducers/states'
 import { Context, interrupt, runInContext } from '../slang'
 import { IVLE_KEY } from '../utils/constants'
